requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
pydantic[email]==2.6.1
pydantic-settings==2.1.0
idna==3.6
certifi==2024.2.2 